moved infinite post list to a separate widget
